+++ UITests/NablaUITestsHostApp/NablaUITestsHostApp/Mocks/PaginatedWatcherMock.swift
@@ -1,12 +1,12 @@
 import NablaMessagingCore
 
 final class PaginatedWatcherMock: PaginatedWatcher {
-    var loadMoreClosure: ((_ completion: @escaping ((Result<Void, Error>) -> Void)) -> Cancellable)?
-    func loadMore(completion: @escaping (Result<Void, Error>) -> Void) -> Cancellable {
+    var loadMoreClosure: ((_ completion: @escaping ((Result<Void, NablaError>) -> Void)) -> Cancellable)?
+    func loadMore(completion: @escaping (Result<Void, NablaError>) -> Void) -> Cancellable {
         loadMoreClosure?(completion) ?? CancellableMock()
     }
 
-    func loadMore(numberOfItems _: Int, completion _: @escaping (Result<Void, Error>) -> Void) -> Cancellable {
+    func loadMore(numberOfItems _: Int, completion _: @escaping (Result<Void, NablaError>) -> Void) -> Cancellable {
         CancellableMock()
     }